+++ src/index.ts
@@ -1,4 +1,2 @@
 export { createGlobalState } from './createGlobalState';
 export { createStore } from './createStore';
-
-export { reduxDevToolsExt } from './devtools';